Major restructuring + test adaptation
