+++ .devcontainer/docker-compose.yml
@@ -17,7 +17,9 @@ services:
         USER_GID: 1000
 
     init: true
-    privileged: true
+
+    extra_hosts:
+      - host.docker.internal:host-gateway
 
     volumes:
       - ..:/workspace:cached
