+++ src/ghga_connector/core/downloading/downloader.py
@@ -12,24 +12,28 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""This file contains all api calls related to uploading files"""
+#
 
 import base64
 import concurrent.futures
 from collections.abc import Iterator, Sequence
 from queue import Queue
 from time import sleep
-from typing import Any, Union
+from typing import Any
 
 import httpx
 
 from ghga_connector.core import exceptions
-from ghga_connector.core.api_calls.work_package import WorkPackageAccessor
-from ghga_connector.core.constants import TIMEOUT, TIMEOUT_LONG
+from ghga_connector.core.api_calls import WorkPackageAccessor
+from ghga_connector.core.constants import TIMEOUT
 from ghga_connector.core.dataclasses import PartRange
-from ghga_connector.core.download.download_handler import (
-    DownloaderBase,
+from ghga_connector.core.downloading.abstract_downloader import DownloaderBase
+from ghga_connector.core.downloading.api_calls import (
+    get_download_url,
+    get_envelope_authorization,
+    get_file_authorization,
+)
+from ghga_connector.core.downloading.request_dataclasses import (
     RetryResponse,
     URLResponse,
 )
@@ -66,7 +70,13 @@ class Downloader(DownloaderBase):
         wait_time = 0
         while wait_time < max_wait_time:
             try:
-                response = self.get_download_url()
+                url_and_headers = get_file_authorization(
+                    file_id=self._file_id,
+                    work_package_accessor=self._work_package_accessor,
+                )
+                response = get_download_url(
+                    client=self._client, url_and_headers=url_and_headers
+                )
             except exceptions.BadResponseCodeError as error:
                 message_display.failure(
                     "The request was invalid and returned a bad HTTP status code."
@@ -77,8 +87,6 @@ class Downloader(DownloaderBase):
                 raise error
 
             if isinstance(response, RetryResponse):
-                # typechecker is a bit weird here. Could skip the else and indentation,
-                # but but then it thinks response for this branch is of type URLResponse
                 retry_time = response.retry_after
                 wait_time += retry_time
                 message_display.display(
@@ -89,6 +97,23 @@ class Downloader(DownloaderBase):
                 return response
 
         raise exceptions.MaxWaitTimeExceededError(max_wait_time=max_wait_time)
+
+    def get_download_urls(self) -> Iterator[URLResponse]:
+        """
+        For a specific multi-part download identified by `file_id`, return an iterator to
+        lazily obtain download URLs.
+        """
+        while True:
+            url_and_headers = get_file_authorization(
+                file_id=self._file_id, work_package_accessor=self._work_package_accessor
+            )
+            url_response = get_download_url(
+                client=self._client, url_and_headers=url_and_headers
+            )
+            if isinstance(url_response, RetryResponse):
+                # File should be staged at that point in time
+                raise exceptions.UnexpectedRetryResponseError()
+            yield url_response
 
     def download_content_range(
         self,
@@ -127,7 +152,7 @@ class Downloader(DownloaderBase):
         part_ranges: Sequence[PartRange],
         queue: Queue[tuple[int, bytes]],
     ) -> None:
-        """Download stuff"""
+        """Download all file parts specified by part_ranges"""
         # Download the parts using a thread pool executor
         executor = concurrent.futures.ThreadPoolExecutor(
             max_workers=max_concurrent_downloads,
@@ -135,7 +160,7 @@ class Downloader(DownloaderBase):
 
         for part_range, download_url in zip(part_ranges, url_response):
             kwargs: dict[str, Any] = {
-                "download_url": download_url,
+                "download_url": download_url.download_url,
                 "start": part_range.start,
                 "end": part_range.stop,
                 "queue": queue,
@@ -143,62 +168,21 @@ class Downloader(DownloaderBase):
 
             executor.submit(self.download_content_range, **kwargs)
 
-    def get_download_url(
-        self,
-    ) -> Union[RetryResponse, URLResponse]:
-        """
-        Perform a RESTful API call to retrieve a presigned download URL.
-        Returns:
-            If the download url is not available yet, a RetryResponse is returned,
-            containing the time in seconds after which the download url should become
-            available.
-            Otherwise, a URLResponse containing the download url and file size in bytes
-            is returned.
-        """
-        return get_download_url(
-            client=self._client,
-            file_id=self._file_id,
-            work_package_accessor=self._work_package_accessor,
-        )
-
-    def get_download_urls(
-        self,
-    ) -> Iterator[URLResponse]:
-        """
-        For a specific multi-part upload identified by the `file_id`, it returns an
-        iterator to obtain download_urls.
-        """
-        while True:
-            url_response = self.get_download_url()
-            if isinstance(url_response, RetryResponse):
-                raise exceptions.UnexcpectedRetryResponseError()
-            yield url_response
-
     def get_file_header_envelope(self) -> bytes:
         """
         Perform a RESTful API call to retrieve a file header envelope.
         Returns:
             The file header envelope (bytes object)
         """
-        # fetch a work order token
-        decrypted_token = self._work_package_accessor.get_work_order_token(
-            file_id=self._file_id
+        url_and_headers = get_envelope_authorization(
+            file_id=self._file_id, work_package_accessor=self._work_package_accessor
         )
-
-        # build url and headers
-        url = f"{self._work_package_accessor.dcs_api_url}/objects/{self._file_id}/envelopes"
-
-        headers = httpx.Headers(
-            {
-                "Accept": "application/json",
-                "Authorization": f"Bearer {decrypted_token}",
-                "Content-Type": "application/json",
-            }
-        )
-
+        url = url_and_headers.endpoint_url
         # Make function call to get download url
         try:
-            response = self._client.get(url=url, headers=headers, timeout=TIMEOUT)
+            response = self._client.get(
+                url=url, headers=url_and_headers.headers, timeout=TIMEOUT
+            )
         except httpx.RequestError as request_error:
             raise exceptions.RequestFailedError(url=url) from request_error
 
@@ -231,72 +215,3 @@ class Downloader(DownloaderBase):
 
         ResponseExceptionTranslator(spec=spec).handle(response=response)
         raise exceptions.BadResponseCodeError(url=url, response_code=status_code)
-
-
-def get_download_url(
-    client: httpx.Client, file_id: str, work_package_accessor: WorkPackageAccessor
-) -> Union[RetryResponse, URLResponse]:
-    """
-    Perform a RESTful API call to retrieve a presigned download URL.
-    Returns:
-        If the download url is not available yet, a RetryResponse is returned,
-        containing the time in seconds after which the download url should become
-        available.
-        Otherwise, a URLResponse containing the download url and file size in bytes
-        is returned.
-    """
-    # fetch a work order token
-    decrypted_token = work_package_accessor.get_work_order_token(file_id=file_id)
-
-    # build url and headers
-    url = f"{work_package_accessor.dcs_api_url}/objects/{file_id}"
-    headers = httpx.Headers(
-        {
-            "Accept": "application/json",
-            "Authorization": f"Bearer {decrypted_token}",
-            "Content-Type": "application/json",
-        }
-    )
-
-    # Make function call to get download url
-    try:
-        response = client.get(url=url, headers=headers, timeout=TIMEOUT_LONG)
-    except httpx.RequestError as request_error:
-        exceptions.raise_if_connection_failed(request_error=request_error, url=url)
-        raise exceptions.RequestFailedError(url=url) from request_error
-
-    status_code = response.status_code
-    if status_code != 200:
-        if status_code == 403:
-            content = response.json()
-            # handle both normal and httpyexpect 403 response
-            if "description" in content:
-                cause = content["description"]
-            else:
-                cause = content["detail"]
-            raise exceptions.UnauthorizedAPICallError(url=url, cause=cause)
-        if status_code != 202:
-            raise exceptions.BadResponseCodeError(url=url, response_code=status_code)
-
-        headers = response.headers
-        if "retry-after" not in headers:
-            raise exceptions.RetryTimeExpectedError(url=url)
-
-        return RetryResponse(retry_after=int(headers["retry-after"]))
-
-    # look for an access method of type s3 in the response:
-    response_body = response.json()
-    download_url = None
-    access_methods = response_body["access_methods"]
-    for access_method in access_methods:
-        if access_method["type"] == "s3":
-            download_url = access_method["access_url"]["url"]
-            file_size = response_body["size"]
-            break
-    else:
-        raise exceptions.NoS3AccessMethodError(url=url)
-
-    return URLResponse(
-        download_url=download_url,
-        file_size=file_size,
-    )
