+++ src/ghga_connector/core/downloading/request_dataclasses.py
@@ -16,10 +16,27 @@
 """Module for reusable wrapper classes improving readability"""
 from dataclasses import dataclass
 
+from httpx import Headers
+
 
 @dataclass
-class PartRange:
-    """Container for inclusive download ranges"""
+class RetryResponse:
+    """Response to download request if file is not yet staged"""
 
-    start: int
-    stop: int
+    retry_after: int
+
+
+@dataclass
+class URLResponse:
+    """Response to download request, containing file size and presigned object storage URL for download"""
+
+    download_url: str
+    file_size: int
+
+
+@dataclass
+class UrlAndHeaders:
+    """Combination of endpoint url and headers needed to make an authorized call against the endpoint"""
+
+    endpoint_url: str
+    headers: Headers
