+++ src/ghga_connector/core/file_operations.py
@@ -16,198 +16,12 @@
 
 """Contains calls of the Presigned URLs in order to Up- and Download Files"""
 
-import base64
-import hashlib
 import math
-import os
-from abc import ABC, abstractmethod
-from collections.abc import Generator, Iterator, Sequence
+from collections.abc import Iterator, Sequence
 from io import BufferedReader
 from pathlib import Path
 
-import crypt4gh.header
-import crypt4gh.keys
-import crypt4gh.lib
-from nacl.bindings import crypto_aead_chacha20poly1305_ietf_encrypt
-
 from ghga_connector.core.dataclasses import PartRange
-
-
-class Checksums:
-    """Container for checksum calculation"""
-
-    def __init__(self):
-        self._unencrypted_sha256 = hashlib.sha256()
-        self._encrypted_md5: list[str] = []
-        self._encrypted_sha256: list[str] = []
-
-    def __repr__(self) -> str:
-        """Return multiline representation of checksum hashes"""
-        return (
-            f"Unencrypted: {self._unencrypted_sha256.hexdigest()}\n"
-            + f"Encrypted MD5: {self._encrypted_md5}\n"
-            + f"Encrypted SHA256: {self._encrypted_sha256}"
-        )
-
-    def encrypted_is_empty(self):
-        """Returns true, the encryption checksum buffer is still empty"""
-        return len(self._encrypted_md5) > 0
-
-    def get(self):
-        """Return all checksums at the end of processing"""
-        return (
-            self._unencrypted_sha256.hexdigest(),
-            self._encrypted_md5,
-            self._encrypted_sha256,
-        )
-
-    def update_unencrypted(self, part: bytes):
-        """Update checksum for unencrypted file"""
-        self._unencrypted_sha256.update(part)
-
-    def update_encrypted(self, part: bytes):
-        """Update encrypted part checksums"""
-        self._encrypted_md5.append(hashlib.md5(part, usedforsecurity=False).hexdigest())
-        self._encrypted_sha256.append(hashlib.sha256(part).hexdigest())
-
-
-class Decryptor(ABC):
-    """Convenience class to deal with file decryption"""
-
-    @abstractmethod
-    def decrypt_file(self, *, input_path: Path, output_path: Path):
-        """Decrypt provided file"""
-
-
-class Crypt4GHDecryptor(Decryptor):
-    """Convenience class to deal with Crypt4GH decryption"""
-
-    def __init__(self, decryption_key_path: Path):
-        self._decryption_key = crypt4gh.keys.get_private_key(
-            decryption_key_path, callback=None
-        )
-
-    def decrypt_file(self, *, input_path: Path, output_path: Path):
-        """Decrypt provided file using Crypt4GH lib"""
-        keys = [(0, self._decryption_key, None)]
-        with input_path.open("rb") as infile, output_path.open("wb") as outfile:
-            crypt4gh.lib.decrypt(keys=keys, infile=infile, outfile=outfile)
-
-
-class Encryptor(ABC):
-    """Handles on the fly encryption and checksum calculation"""
-
-    @abstractmethod
-    def get_encrypted_size(self) -> int:
-        """Get file size after encryption, excluding envelope"""
-
-    @abstractmethod
-    def process_file(self, file: BufferedReader) -> Generator[bytes, None, None]:
-        """Encrypt file parts and prepare for upload."""
-
-
-class Crypt4GHEncryptor(Encryptor):
-    """Handles on the fly encryption and checksum calculation"""
-
-    def __init__(  # noqa: PLR0913
-        self,
-        part_size: int,
-        private_key_path: Path,
-        server_public_key: str,
-        checksums: Checksums = Checksums(),
-        file_secret: bytes = os.urandom(32),
-    ):
-        self._encrypted_file_size = 0
-        self._checksums = checksums
-        self._file_secret = file_secret
-        self._part_size = part_size
-        self._private_key_path = private_key_path
-        self._server_public_key = base64.b64decode(server_public_key)
-
-    def _encrypt(self, part: bytes):
-        """Encrypt file part using secret"""
-        segments, incomplete_segment = get_segments(
-            part=part, segment_size=crypt4gh.lib.SEGMENT_SIZE
-        )
-
-        encrypted_segments = []
-        for segment in segments:
-            encrypted_segments.append(self._encrypt_segment(segment))
-
-        return b"".join(encrypted_segments), incomplete_segment
-
-    def _encrypt_segment(self, segment: bytes):
-        """Encrypt one single segment"""
-        nonce = os.urandom(12)
-        encrypted_data = crypto_aead_chacha20poly1305_ietf_encrypt(
-            segment, None, nonce, self._file_secret
-        )  # no aad
-        return nonce + encrypted_data
-
-    def _create_envelope(self) -> bytes:
-        """
-        Gather file encryption/decryption secret and assemble a crypt4gh envelope using the
-        servers private and the clients public key
-        """
-        private_key = crypt4gh.keys.get_private_key(
-            self._private_key_path, callback=None
-        )
-        keys = [(0, private_key, self._server_public_key)]
-        header_content = crypt4gh.header.make_packet_data_enc(0, self._file_secret)
-        header_packets = crypt4gh.header.encrypt(header_content, keys)
-        header_bytes = crypt4gh.header.serialize(header_packets)
-
-        return header_bytes
-
-    def get_encrypted_size(self) -> int:
-        """Get file size after encryption, excluding envelope"""
-        return self._encrypted_file_size
-
-    def process_file(self, file: BufferedReader) -> Generator[bytes, None, None]:
-        """Encrypt file parts and prepare for upload."""
-        unprocessed_bytes = b""
-        upload_buffer = self._create_envelope()
-
-        # get envelope size to adjust checksum buffers and encrypted content size
-        envelope_size = len(upload_buffer)
-
-        for file_part in read_file_parts(file=file, part_size=self._part_size):
-            # process unencrypted
-            self._checksums.update_unencrypted(file_part)
-            unprocessed_bytes += file_part
-
-            # encrypt in chunks
-            encrypted_bytes, unprocessed_bytes = self._encrypt(unprocessed_bytes)
-            upload_buffer += encrypted_bytes
-
-            # update checksums and yield if part size
-            if len(upload_buffer) >= self._part_size:
-                current_part = upload_buffer[: self._part_size]
-                if self._checksums.encrypted_is_empty():
-                    self._checksums.update_encrypted(current_part[envelope_size:])
-                else:
-                    self._checksums.update_encrypted(current_part)
-                self._encrypted_file_size += self._part_size
-                yield current_part
-                upload_buffer = upload_buffer[self._part_size :]
-
-        # process dangling bytes
-        if unprocessed_bytes:
-            upload_buffer += self._encrypt_segment(unprocessed_bytes)
-
-        while len(upload_buffer) >= self._part_size:
-            current_part = upload_buffer[: self._part_size]
-            self._checksums.update_encrypted(current_part)
-            self._encrypted_file_size += self._part_size
-            yield current_part
-            upload_buffer = upload_buffer[self._part_size :]
-
-        if upload_buffer:
-            self._checksums.update_encrypted(upload_buffer)
-            self._encrypted_file_size += len(upload_buffer)
-            yield upload_buffer
-
-        self._encrypted_file_size -= envelope_size
 
 
 def is_file_encrypted(file_path: Path):
