+++ src/ghga_connector/core/uploading/__init__.py
@@ -13,13 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""Module for reusable wrapper classes improving readability"""
-from dataclasses import dataclass
+"""TODO"""
 
 
-@dataclass
-class PartRange:
-    """Container for inclusive download ranges"""
-
-    start: int
-    stop: int
+from .api_calls import Uploader  # noqa: F401
+from .main import run_upload  # noqa: F401
+from .uploader import UploaderBase  # noqa: F401
